#50 refactor: intent 参照元を新 _stores/intent/ へ切替
- hook (useIntentStore) は新 _stores/intent/ (実装/context) を参照
- stores-provider.tsx は createIntentStore/IntentStoreContext を新 _stores/intent/ から取得
  (旧 _stores/intent-store.ts・_contexts/intent-store-context.tsx 自体はまだ削除しない)
