+++ src/prototypes/time-control/time-control-03/_components/actor-controller/index.tsx
@@ -4,11 +4,11 @@ import { Button } from '@/components/ui/button'
 
 import { formatCoord } from '../../../time-control-02/_lib/format-coord'
 import { getTickMs } from '../../../time-control-02/_lib/get-tick-ms'
-import { useIntentStore } from '../../_contexts/intent-store-context'
 import { useActorStore } from '../../_stores/actor'
 import { DEFAULT_ACTOR_INFO } from '../../_stores/actor/constants'
 import { useActorSettingsStore } from '../../_stores/actor-settings'
 import { DEFAULT_ACTOR_SETTINGS } from '../../_stores/actor-settings/constants'
+import { useIntentStore } from '../../_stores/intent'
 import { usePathStore } from '../../_stores/path'
 import { DEFAULT_PATH } from '../../_stores/path/constants'
 import { usePositionStore } from '../../_stores/position'
